+++ src/integrationTest/java/org/opensearch/security/systemindex/sampleplugin/RestIndexDocumentIntoSystemIndexAction.java
@@ -8,7 +8,7 @@
  *
  */
 
-package org.opensearch.security.plugin;
+package org.opensearch.security.systemindex.sampleplugin;
 
 import java.util.List;
 
@@ -17,35 +17,33 @@ import org.opensearch.client.node.NodeClient;
 import org.opensearch.rest.BaseRestHandler;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.rest.action.RestToXContentListener;
-import org.opensearch.security.identity.PluginContextSwitcher;
 
 import static java.util.Collections.singletonList;
-import static org.opensearch.rest.RestRequest.Method.GET;
+import static org.opensearch.rest.RestRequest.Method.PUT;
 
-public class RestRunClusterHealthAction extends BaseRestHandler {
+public class RestIndexDocumentIntoSystemIndexAction extends BaseRestHandler {
 
     private final Client client;
-    private final PluginContextSwitcher contextSwitcher;
 
-    public RestRunClusterHealthAction(Client client, PluginContextSwitcher contextSwitcher) {
+    public RestIndexDocumentIntoSystemIndexAction(Client client) {
         this.client = client;
-        this.contextSwitcher = contextSwitcher;
     }
 
     @Override
     public List<Route> routes() {
-        return singletonList(new Route(GET, "/try-cluster-health/{runAs}"));
+        return singletonList(new Route(PUT, "/try-create-and-index/{index}"));
     }
 
     @Override
     public String getName() {
-        return "test_run_cluster_health_action";
+        return "test_index_document_into_system_index_action";
     }
 
     @Override
     public RestChannelConsumer prepareRequest(RestRequest request, NodeClient client) {
         String runAs = request.param("runAs");
-        RunClusterHealthRequest runRequest = new RunClusterHealthRequest(runAs);
-        return channel -> client.execute(RunClusterHealthAction.INSTANCE, runRequest, new RestToXContentListener<>(channel));
+        String indexName = request.param("index");
+        IndexDocumentIntoSystemIndexRequest indexRequest = new IndexDocumentIntoSystemIndexRequest(indexName, runAs);
+        return channel -> client.execute(IndexDocumentIntoSystemIndexAction.INSTANCE, indexRequest, new RestToXContentListener<>(channel));
     }
 }
