+++ src/integrationTest/java/org/opensearch/security/systemindex/sampleplugin/RunClusterHealthAction.java
@@ -8,7 +8,7 @@
  *
  */
 
-package org.opensearch.security.plugin;
+package org.opensearch.security.systemindex.sampleplugin;
 
 import org.opensearch.action.ActionType;
