+++ src/integrationTest/java/org/opensearch/security/systemindex/sampleplugin/RunClusterHealthRequest.java
@@ -8,7 +8,7 @@
  *
  */
 
-package org.opensearch.security.plugin;
+package org.opensearch.security.systemindex.sampleplugin;
 
 import java.io.IOException;
